backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
python-dotenv==1.0.1
websockets==13.1
requests==2.32.3
beautifulsoup4==4.12.3
feedparser==6.0.11
alpaca-py==0.33.1
numpy==2.1.0
schedule==1.2.2
python-telegram-bot==21.6
